spenzy-expense-service/requirements.txt
grpcio>=1.68.0
grpcio-tools>=1.68.0
grpcio-reflection>=1.68.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
greenlet>=3.0.0
-e ../spenzy-common 